+++ app/controllers/internal_controller.rb
@@ -14,7 +14,7 @@ class InternalController < ApplicationController
 
   # GET /directory
   def directory
-    @users = User.current.where(approved: true).order(id: :desc).limit(5)
+    @users = User.current.where(approved: true).order(id: :desc).limit(20)
     @key_contacts = User.current.where(key_contact: true)
   end